+++ otx/algorithms/segmentation/adapters/mmseg/task.py
@@ -1,4 +1,4 @@
-"""Task of OTX Classification using mmclassification training backend."""
+"""Task of OTX Segmentation using mmsegmentation training backend."""
 
 # Copyright (C) 2023 Intel Corporation
 #
@@ -15,6 +15,7 @@
 # and limitations under the License.
 
 import glob
+import io
 import os
 import time
 from contextlib import nullcontext
@@ -22,32 +23,23 @@ from copy import deepcopy
 from typing import Any, Dict, Optional, Union
 
 import torch
-from mmcls.apis import train_model
-from mmcls.datasets import build_dataloader, build_dataset
-from mmcls.utils import collect_env
 from mmcv.runner import wrap_fp16_model
-from mmcv.utils import Config, ConfigDict
+from mmcv.utils import Config, ConfigDict, get_git_hash
+from mmseg import __version__
+from mmseg.apis import train_segmentor
+from mmseg.datasets import build_dataloader, build_dataset
+from mmseg.utils import collect_env
 
-from otx.algorithms.classification.adapters.mmcls.utils.exporter import (
-    ClassificationExporter,
-)
-from otx.algorithms.classification.task import OTXClassificationTask
 from otx.algorithms.common.adapters.mmcv.hooks.recording_forward_hook import (
-    ActivationMapHook,
     BaseRecordingForwardHook,
-    EigenCamHook,
     FeatureVectorHook,
-    ReciproCAMHook,
 )
 from otx.algorithms.common.adapters.mmcv.utils import (
     build_data_parallel,
     get_configs_by_pairs,
     patch_data_pipeline,
+    patch_from_hyperparams,
 )
-from otx.algorithms.common.adapters.mmcv.utils import (
-    build_dataloader as otx_build_dataloader,
-)
-from otx.algorithms.common.adapters.mmcv.utils import build_dataset as otx_build_dataset
 from otx.algorithms.common.adapters.mmcv.utils.config_utils import (
     MPAConfig,
     update_or_add_custom_hook,
@@ -56,20 +48,28 @@ from otx.algorithms.common.configs.training_base import TrainType
 from otx.algorithms.common.utils import set_random_seed
 from otx.algorithms.common.utils.data import get_dataset
 from otx.algorithms.common.utils.logger import get_logger
+from otx.algorithms.segmentation.adapters.mmseg.configurer import (
+    IncrSegmentationConfigurer,
+    SegmentationConfigurer,
+    SemiSLSegmentationConfigurer,
+)
+from otx.algorithms.segmentation.adapters.mmseg.utils.builder import build_segmentor
+from otx.algorithms.segmentation.adapters.mmseg.utils.exporter import SegmentationExporter
+from otx.algorithms.segmentation.task import OTXSegmentationTask
+
+# from otx.algorithms.segmentation.utils import get_det_model_api_configuration
+from otx.api.configuration import cfg_helper
+from otx.api.configuration.helper.utils import ids_to_strings
 from otx.api.entities.datasets import DatasetEntity
-from otx.api.entities.explain_parameters import ExplainParameters
 from otx.api.entities.inference_parameters import InferenceParameters
-from otx.api.entities.model import ModelPrecision
+from otx.api.entities.model import (
+    ModelEntity,
+    ModelPrecision,
+)
 from otx.api.entities.subset import Subset
 from otx.api.entities.task_environment import TaskEnvironment
+from otx.api.serialization.label_mapper import label_schema_to_bytes
 from otx.core.data import caching
-
-from .configurer import (
-    ClassificationConfigurer,
-    IncrClassificationConfigurer,
-    SemiSLClassificationConfigurer,
-)
-from .utils import build_classifier
 
 logger = get_logger()
 
@@ -77,8 +77,8 @@ logger = get_logger()
 # pylint: disable=too-many-lines
 
 
-class MMClassificationTask(OTXClassificationTask):
-    """Task class for OTX classification using mmclassification training backend."""
+class MMSegmentationTask(OTXSegmentationTask):
+    """Task class for OTX segmentation using mmsegmentation training backend."""
 
     # pylint: disable=too-many-instance-attributes
     def __init__(self, task_environment: TaskEnvironment, output_path: Optional[str] = None):
@@ -89,19 +89,8 @@ class MMClassificationTask(OTXClassificationTask):
     # pylint: disable=too-many-locals, too-many-branches, too-many-statements
     def _init_task(self, export: bool = False):  # noqa
         """Initialize task."""
-
-        if self._multilabel:
-            cfg_path = os.path.join(self._model_dir, "model_multilabel.py")
-        elif self._hierarchical:
-            cfg_path = os.path.join(self._model_dir, "model_hierarchical.py")
-        else:
-            cfg_path = os.path.join(self._model_dir, "model.py")
-        self._recipe_cfg = MPAConfig.fromfile(cfg_path)
+        self._recipe_cfg = MPAConfig.fromfile(os.path.join(self._model_dir, "model.py"))
         self._recipe_cfg.domain = self._task_type.domain
-        self._recipe_cfg.model.multilabel = self._multilabel
-        self._recipe_cfg.model.hierarchical = self._hierarchical
-        if self._hierarchical:
-            self._recipe_cfg.model.head.hierarchical_info = self._hierarchical_info
         self._config = self._recipe_cfg
 
         set_random_seed(self._recipe_cfg.get("seed", 5), logger, self._recipe_cfg.get("deterministic", False))
@@ -110,7 +99,7 @@ class MMClassificationTask(OTXClassificationTask):
         patch_data_pipeline(self._recipe_cfg, self.data_pipeline_path)
 
         if not export:
-            self._recipe_cfg.merge_from_dict(self._init_hparam())
+            patch_from_hyperparams(self._recipe_cfg, self._hyperparams)
 
         if "custom_hooks" in self.override_configs:
             override_custom_hooks = self.override_configs.pop("custom_hooks")
@@ -141,6 +130,8 @@ class MMClassificationTask(OTXClassificationTask):
         # Update recipe with caching modules
         self._update_caching_modules(self._recipe_cfg.data)
 
+        logger.info("initialized.")
+
     # pylint: disable=too-many-arguments
     def configure(
         self,
@@ -148,7 +139,7 @@ class MMClassificationTask(OTXClassificationTask):
         subset="train",
         ir_options=None,
     ):
-        """Patch mmcv configs for OTX classification settings."""
+        """Patch mmcv configs for OTX segmentation settings."""
 
         # deepcopy all configs to make sure
         # changes under MPA and below does not take an effect to OTX for clear distinction
@@ -166,35 +157,13 @@ class MMClassificationTask(OTXClassificationTask):
         recipe_cfg.resume = self._resume
 
         if self._train_type == TrainType.Incremental:
-            configurer = IncrClassificationConfigurer()
+            configurer = IncrSegmentationConfigurer()
         elif self._train_type == TrainType.Semisupervised:
-            configurer = SemiSLClassificationConfigurer()
+            configurer = SemiSLSegmentationConfigurer()
         else:
-            configurer = ClassificationConfigurer()
-
-        options_for_patch_datasets = {"type": "OTXClsDataset", "empty_label": self._empty_label}
-        options_for_patch_evaluation = {"task": "normal"}
-        if self._multilabel:
-            options_for_patch_datasets["type"] = "OTXMultilabelClsDataset"
-            options_for_patch_evaluation["task"] = "multilabel"
-        elif self._hierarchical:
-            options_for_patch_datasets["type"] = "OTXHierarchicalClsDataset"
-            options_for_patch_datasets["hierarchical_info"] = self._hierarchical_info
-            options_for_patch_evaluation["task"] = "hierarchical"
-        elif self._selfsl:
-            options_for_patch_datasets["type"] = "SelfSLDataset"
-
+            configurer = SegmentationConfigurer()
         cfg = configurer.configure(
-            recipe_cfg,
-            self._model_ckpt,
-            data_cfg,
-            training,
-            subset,
-            ir_options,
-            data_classes,
-            model_classes,
-            options_for_patch_datasets=options_for_patch_datasets,
-            options_for_patch_evaluation=options_for_patch_evaluation,
+            recipe_cfg, self._model_ckpt, data_cfg, training, subset, ir_options, data_classes, model_classes
         )
         self._config = cfg
         return cfg
@@ -206,7 +175,7 @@ class MMClassificationTask(OTXClassificationTask):
         **kwargs,
     ) -> torch.nn.Module:
         """Build model from model_builder."""
-        model_builder = getattr(self, "model_builder", build_classifier)
+        model_builder = getattr(self, "model_builder", build_segmentor)
         model = model_builder(cfg, **kwargs)
         if bool(fp16):
             wrap_fp16_model(model)
@@ -232,12 +201,15 @@ class MMClassificationTask(OTXClassificationTask):
         )
 
         dump_features = True
-        dump_saliency_map = not inference_parameters.is_evaluation if inference_parameters else True
 
         self._init_task()
 
         cfg = self.configure(False, "test", None)
         logger.info("infer!")
+
+        # FIXME: Currently segmentor does not support multi batch inference.
+        if "test" in cfg.data and "test_dataloader" in cfg.data:
+            cfg.data.test_dataloader["samples_per_gpu"] = 1
 
         # Data loader
         mm_dataset = build_dataset(cfg.data.test)
@@ -248,13 +220,26 @@ class MMClassificationTask(OTXClassificationTask):
             num_gpus=len(cfg.gpu_ids),
             dist=cfg.distributed,
             seed=cfg.get("seed", None),
+            persistent_workers=False,
             shuffle=False,
         )
 
+        # Target classes
+        if "task_adapt" in cfg:
+            target_classes = cfg.task_adapt.final
+            if len(target_classes) < 1:
+                raise KeyError(
+                    f"target_classes={target_classes} is empty check the metadata from model ckpt or recipe "
+                    "configuration"
+                )
+        else:
+            target_classes = mm_dataset.CLASSES
+
         # Model
         model = self.build_model(cfg, fp16=cfg.get("fp16", False))
+        model.CLASSES = target_classes
         model.eval()
-        feature_model = model
+        feature_model = model.model_s if self._train_type == TrainType.Semisupervised else model
         model = build_data_parallel(model, cfg, distributed=False)
 
         # InferenceProgressCallback (Time Monitor enable into Infer task)
@@ -274,45 +259,33 @@ class MMClassificationTask(OTXClassificationTask):
             model.register_forward_pre_hook(pre_hook)
             model.register_forward_hook(hook)
 
-        if not dump_saliency_map:
-            forward_explainer_hook: Union[nullcontext, BaseRecordingForwardHook] = nullcontext()
-        else:
-            forward_explainer_hook = ReciproCAMHook(feature_model)
+        eval_predictions = []
+        feature_vectors = []
+
         if not dump_features:
             feature_vector_hook: Union[nullcontext, BaseRecordingForwardHook] = nullcontext()
         else:
             feature_vector_hook = FeatureVectorHook(feature_model)
 
-        eval_predictions = []
-        feature_vectors = []
-        saliency_maps = []
-
         with feature_vector_hook:
-            with forward_explainer_hook:
-                for data in dataloader:
-                    with torch.no_grad():
-                        result = model(return_loss=False, **data)
-                    eval_predictions.extend(result)
+            for data in dataloader:
+                with torch.no_grad():
+                    result = model(return_loss=False, output_logits=True, **data)
+                eval_predictions.append(result)
                 if isinstance(feature_vector_hook, nullcontext):
                     feature_vectors = [None] * len(mm_dataset)
                 else:
-                    feature_vectors = feature_vector_hook.records  # pylint: disable=no-member
-                if isinstance(forward_explainer_hook, nullcontext):
-                    saliency_maps = [None] * len(mm_dataset)
-                else:
-                    saliency_maps = forward_explainer_hook.records  # pylint: disable=no-member
-                if len(eval_predictions) == 0:
-                    eval_predictions = [None] * len(mm_dataset)
+                    feature_vectors = feature_vector_hook.records
 
-        assert len(eval_predictions) == len(feature_vectors) == len(saliency_maps), (
-            "Number of elements should be the same, however, number of outputs are "
-            f"{len(eval_predictions)}, {len(feature_vectors)}, and {len(saliency_maps)}"
+        assert len(eval_predictions) == len(feature_vectors), (
+            "Number of elements should be the same, however, number of outputs are ",
+            f"{len(eval_predictions)} and {len(feature_vectors)}",
         )
 
         outputs = dict(
+            classes=target_classes,
             eval_predictions=eval_predictions,
             feature_vectors=feature_vectors,
-            saliency_maps=saliency_maps,
         )
         return outputs
 
@@ -321,7 +294,8 @@ class MMClassificationTask(OTXClassificationTask):
         self,
         dataset: DatasetEntity,
     ):
-        """Train function in MMClassificationTask."""
+        """Train function in MMSegmentationTask."""
+        logger.info("init data cfg.")
         self._data_cfg = ConfigDict(data=ConfigDict())
 
         for cfg_key, subset in zip(
@@ -354,15 +328,33 @@ class MMClassificationTask(OTXClassificationTask):
         # Data
         datasets = [build_dataset(cfg.data.train)]
 
+        # FIXME: Currently segmentor does not support multi batch evaluation.
+        # For the Self-SL case, there is no val data. So, need to check the
+
+        if "val" in cfg.data and "val_dataloader" in cfg.data:
+            cfg.data.val_dataloader["samples_per_gpu"] = 1
+
+        # Target classes
+        if "task_adapt" in cfg:
+            target_classes = cfg.task_adapt.final
+        else:
+            target_classes = datasets[0].CLASSES
+
         # Metadata
         meta = dict()
         meta["env_info"] = env_info
-        meta["seed"] = cfg.get("seed", 5)
+        meta["seed"] = cfg.seed
         meta["exp_name"] = cfg.work_dir
+        if cfg.checkpoint_config is not None:
+            cfg.checkpoint_config.meta = dict(
+                mmseg_version=__version__ + get_git_hash()[:7],
+                CLASSES=target_classes,
+            )
 
         # Model
         model = self.build_model(cfg, fp16=cfg.get("fp16", False))
         model.train()
+        model.CLASSES = target_classes
 
         if cfg.distributed:
             torch.nn.SyncBatchNorm.convert_sync_batchnorm(model)
@@ -375,156 +367,40 @@ class MMClassificationTask(OTXClassificationTask):
                 cfg.optimizer.lr = new_lr
 
         validate = bool(cfg.data.get("val", None))
-        if validate:
-            val_dataset = build_dataset(cfg.data.val, dict(test_mode=True))
-            val_loader_cfg = {
-                "num_gpus": len(cfg.gpu_ids),
-                "dist": cfg.distributed,
-                "round_up": True,
-                "seed": cfg.seed,
-                "shuffle": False,  # Not shuffle by default
-                "sampler_cfg": None,  # Not use sampler by default
-                **cfg.data.get("val_dataloader", {}),
-            }
-            val_dataloader = build_dataloader(val_dataset, **val_loader_cfg)
-            eval_cfg = cfg.get("evaluation", {})
-            eval_cfg["by_epoch"] = cfg.runner["type"] != "IterBasedRunner"
-            cfg.custom_hooks.append(
-                dict(
-                    type="DistCustomEvalHook" if cfg.distributed else "CustomEvalHook",
-                    dataloader=val_dataloader,
-                    priority="ABOVE_NORMAL",
-                    **eval_cfg,
-                )
-            )
-
-        train_model(
+        train_segmentor(
             model,
             datasets,
             cfg,
             distributed=cfg.distributed,
-            validate=False,  # For using CustomEvalHook
+            validate=validate,
             timestamp=timestamp,
             meta=meta,
         )
 
         # Save outputs
         output_ckpt_path = os.path.join(cfg.work_dir, "latest.pth")
-        best_ckpt_path = glob.glob(os.path.join(cfg.work_dir, "best_*.pth"))
+        best_ckpt_path = glob.glob(os.path.join(cfg.work_dir, "best_mDice_*.pth"))
+        if len(best_ckpt_path) > 0:
+            output_ckpt_path = best_ckpt_path[0]
+        best_ckpt_path = glob.glob(os.path.join(cfg.work_dir, "best_mIoU_*.pth"))
         if len(best_ckpt_path) > 0:
             output_ckpt_path = best_ckpt_path[0]
         return dict(
             final_ckpt=output_ckpt_path,
         )
 
-    # These need to be moved somewhere
-    def _update_caching_modules(self, data_cfg: Config) -> None:
-        def _find_max_num_workers(cfg: dict):
-            num_workers = [0]
-            for key, value in cfg.items():
-                if key == "workers_per_gpu" and isinstance(value, int):
-                    num_workers += [value]
-                elif isinstance(value, dict):
-                    num_workers += [_find_max_num_workers(value)]
+    def _explain_model(self):
+        """Explain function of OTX Segmentation Task."""
+        raise NotImplementedError
 
-            return max(num_workers)
-
-        def _get_mem_cache_size():
-            if not hasattr(self._hyperparams.algo_backend, "mem_cache_size"):
-                return 0
-
-            return self._hyperparams.algo_backend.mem_cache_size
-
-        max_num_workers = _find_max_num_workers(data_cfg)
-        mem_cache_size = _get_mem_cache_size()
-
-        mode = "multiprocessing" if max_num_workers > 0 else "singleprocessing"
-        caching.MemCacheHandlerSingleton.create(mode, mem_cache_size)
-
-        update_or_add_custom_hook(
-            self._recipe_cfg,
-            ConfigDict(type="MemCacheHook", priority="VERY_LOW"),
-        )
-
-    def _explain_model(self, dataset: DatasetEntity, explain_parameters: Optional[ExplainParameters]):
-        """Explain function in MMClassificationTask."""
-        explainer_hook_selector = {
-            "eigencam": EigenCamHook,
-            "activationmap": ActivationMapHook,
-            "classwisesaliencymap": ReciproCAMHook,
-        }
-
-        self._data_cfg = ConfigDict(
-            data=ConfigDict(
-                train=ConfigDict(
-                    otx_dataset=None,
-                    labels=self._labels,
-                ),
-                test=ConfigDict(
-                    otx_dataset=dataset,
-                    labels=self._labels,
-                ),
-            )
-        )
-
-        self._init_task()
-        cfg = self.configure(False, "test", None)
-        # Data loader
-        mm_dataset = otx_build_dataset(cfg, "test", build_dataset)
-        dataloader = otx_build_dataloader(
-            mm_dataset,
-            cfg,
-            "test",
-            build_dataloader,
-            distributed=False,
-            round_up=False,
-        )
-        model = self.build_model(cfg, fp16=cfg.get("fp16", False))
-        model.eval()
-        feature_model = model
-        model = build_data_parallel(model, cfg, distributed=False)
-
-        # InferenceProgressCallback (Time Monitor enable into Infer task)
-        time_monitor = None
-        if cfg.get("custom_hooks", None):
-            time_monitor = [hook.time_monitor for hook in cfg.custom_hooks if hook.type == "OTXProgressHook"]
-            time_monitor = time_monitor[0] if time_monitor else None
-        if time_monitor is not None:
-
-            def pre_hook(module, inp):  # pylint: disable=unused-argument
-                time_monitor.on_test_batch_begin(None, None)
-
-            def hook(module, inp, outp):  # pylint: disable=unused-argument
-                time_monitor.on_test_batch_end(None, None)
-
-            model.register_forward_pre_hook(pre_hook)
-            model.register_forward_hook(hook)
-
-        explainer = explain_parameters.explainer if explain_parameters else None
-        if explainer is not None:
-            explainer_hook = explainer_hook_selector.get(explainer.lower(), None)
-        else:
-            explainer_hook = None
-        if explainer_hook is None:
-            raise NotImplementedError("Explainer algorithm not supported!")
-
-        eval_predictions = []
-        with explainer_hook(feature_model) as forward_explainer_hook:
-            # do inference and record intermediate fmap
-            for data in dataloader:
-                with torch.no_grad():
-                    result = model(return_loss=False, **data)
-                eval_predictions.extend(result)
-            saliency_maps = forward_explainer_hook.records
-
-        assert len(eval_predictions) == len(saliency_maps), (
-            "Number of elements should be the same, however, number of outputs are "
-            f"{len(eval_predictions)}, and {len(saliency_maps)}"
-        )
-
-        return eval_predictions, saliency_maps
-
-    def _export_model(self, precision, dump_features):
+    # pylint: disable=too-many-statements
+    def _export_model(
+        self,
+        precision: ModelPrecision = ModelPrecision.FP32,
+        dump_features: bool = True,
+    ):
+        """Export function of OTX Segmentation Task."""
+        # copied from OTX inference_task.py
         self._init_task(export=True)
 
         cfg = self.configure(False, "test", None)
@@ -544,18 +420,19 @@ class MMClassificationTask(OTXClassificationTask):
             if export_options["deploy_cfg"]["codebase_config"]["task"] != "Segmentation":
                 if "saliency_map" not in output_names:
                     output_names.append("saliency_map")
-        export_options["model_builder"] = getattr(self, "model_builder", build_classifier)
+        export_options["model_builder"] = getattr(self, "model_builder", build_segmentor)
 
         if self._precision[0] == ModelPrecision.FP16:
             export_options["deploy_cfg"]["backend_config"]["mo_options"]["flags"].append("--compress_to_fp16")
 
-        exporter = ClassificationExporter()
+        exporter = SegmentationExporter()
         results = exporter.run(
             cfg,
             **export_options,
         )
         return results
 
+    # This should moved somewhere
     def _init_deploy_cfg(self) -> Union[Config, None]:
         base_dir = os.path.abspath(os.path.dirname(self._task_environment.model_template.model_template_path))
         deploy_cfg_path = os.path.join(base_dir, "deployment.py")
@@ -627,57 +504,60 @@ class MMClassificationTask(OTXClassificationTask):
 
         return deploy_cfg
 
-    def _init_hparam(self) -> dict:
-        params = self._hyperparams.learning_parameters
-        warmup_iters = int(params.learning_rate_warmup_iters)
-        if self._multilabel:
-            # hack to use 1cycle policy
-            lr_config = ConfigDict(max_lr=params.learning_rate, warmup=None)
-        else:
-            lr_config = (
-                ConfigDict(warmup_iters=warmup_iters) if warmup_iters > 0 else ConfigDict(warmup_iters=0, warmup=None)
-            )
-
-        early_stop = False
-        if self._recipe_cfg is not None:
-            if params.enable_early_stopping and self._recipe_cfg.get("evaluation", None):
-                early_stop = ConfigDict(
-                    start=int(params.early_stop_start),
-                    patience=int(params.early_stop_patience),
-                    iteration_patience=int(params.early_stop_iteration_patience),
-                )
-
-        if self._recipe_cfg.runner.get("type").startswith("IterBasedRunner"):  # type: ignore
-            runner = ConfigDict(max_iters=int(params.num_iters))
-        else:
-            runner = ConfigDict(max_epochs=int(params.num_iters))
-
-        config = ConfigDict(
-            optimizer=ConfigDict(lr=params.learning_rate),
-            lr_config=lr_config,
-            early_stop=early_stop,
-            data=ConfigDict(
-                samples_per_gpu=int(params.batch_size),
-                workers_per_gpu=int(params.num_workers),
-            ),
-            runner=runner,
-        )
-
-        if self._train_type.value == "Semisupervised":
-            unlabeled_config = ConfigDict(
-                data=ConfigDict(
-                    unlabeled_dataloader=ConfigDict(
-                        samples_per_gpu=int(params.unlabeled_batch_size),
-                        workers_per_gpu=int(params.num_workers),
-                    )
-                )
-            )
-            config.update(unlabeled_config)
-        return config
-
     # This should be removed
     def update_override_configurations(self, config):
         """Update override_configs."""
         logger.info(f"update override config with: {config}")
         config = ConfigDict(**config)
         self.override_configs.update(config)
+
+    def save_model(self, output_model: ModelEntity):
+        """Save best model weights in SegmentationTrainTask."""
+        logger.info("called save_model")
+        buffer = io.BytesIO()
+        hyperparams_str = ids_to_strings(cfg_helper.convert(self._hyperparams, dict, enum_to_str=True))
+        labels = {label.name: label.color.rgb_tuple for label in self._labels}
+        model_ckpt = torch.load(self._model_ckpt)
+        modelinfo = {
+            "model": model_ckpt,
+            "config": hyperparams_str,
+            "labels": labels,
+            "VERSION": 1,
+        }
+
+        torch.save(modelinfo, buffer)
+        output_model.set_data("weights.pth", buffer.getvalue())
+        output_model.set_data(
+            "label_schema.json",
+            label_schema_to_bytes(self._task_environment.label_schema),
+        )
+        output_model.precision = self._precision
+
+    # These need to be moved somewhere
+    def _update_caching_modules(self, data_cfg: Config) -> None:
+        def _find_max_num_workers(cfg: dict):
+            num_workers = [0]
+            for key, value in cfg.items():
+                if key == "workers_per_gpu" and isinstance(value, int):
+                    num_workers += [value]
+                elif isinstance(value, dict):
+                    num_workers += [_find_max_num_workers(value)]
+
+            return max(num_workers)
+
+        def _get_mem_cache_size():
+            if not hasattr(self._hyperparams.algo_backend, "mem_cache_size"):
+                return 0
+
+            return self._hyperparams.algo_backend.mem_cache_size
+
+        max_num_workers = _find_max_num_workers(data_cfg)
+        mem_cache_size = _get_mem_cache_size()
+
+        mode = "multiprocessing" if max_num_workers > 0 else "singleprocessing"
+        caching.MemCacheHandlerSingleton.create(mode, mem_cache_size)
+
+        update_or_add_custom_hook(
+            self._recipe_cfg,
+            ConfigDict(type="MemCacheHook", priority="VERY_LOW"),
+        )